chore:added schema and fixed model
